backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
psycopg2-binary
asyncpg
pydantic
pydantic-settings
alembic
python-jose[cryptography]
bcrypt
openai
anthropic
google-genai
langchain
python-dotenv
python-dateutil
httpx
cachetools
celery
redis
pytest
pytest-asyncio
httpx
aiosqlite
email-validator
